Initial commit of ttet program.
At this point it's mostly still the example from ggez
that I've taken apart and pulled out some ideas I think
I'll want to use later on.  However as I dig into this,
I'm wondering if it might be better to not make use of
all the fancy drawing and rotation code that ggez gives
us, and instead just make a 2d array game board, then
have a function that draws whatever is set in that array.

It might make collision detection easier.