More work in progress.
Decided to get rid of the rotation and make my own board grid
with a local array that will indicate what to display.  About
to re-work all the drawing logic, but it might be good to keep
some of this around in the event I need to see what I was thinking
of before.