More work on a 2d array for board state.
Updated the board array to be wider.
Improved on the grid drawing.
Added code for tetrominoe I with a move l,r,d function
that will check for collisions.

It's going to be a bunch of very similar code for each tet
when deciding if we can move or rotate, I wonder if there is
a better way to do it..

Gave the window a title.