Finished plotting and movement for all pieces
Finised all the supported Tetrominos with basic movement, rotation,
and prevention of hitting the walls and floor.

Rotations can still carry us into unsupported regions.

It is becoming clear this is a case of "Months of programming can
save you hours of planning" but is that not the whole point of this
exercise?  Now, don't tell me "Perfect practice makes perfect",
because I don't want to hear that.