Added rotation check to piece.
Made sure the next rotation is legal for a given board, piece, and
current rotation.

Note that I will probably have to come back at some point and add
the "bump" rotation, where if the next rotation is invalid, but
the following one is, then skip ahead to the valid rotation.