Updated plot_tet to take the tet type to plot.
By passing the tet type as an arg, we can select Blank or Base
as options and have the board array updated for either.  This allows
me to re-use plot_tet when a piece gets to the bottom and needs
to be transitioned from active to part of the playing field.