Larger board, filled row delete, random piece
Made the board array 4 rows longer to account for four rows that
will not be visible, but are part of the squares where the next
piece is spawned.  Eventually I will not show those squares.

Added (from SO) a random tet selector.

New code and board type for a filled row.  We now delete a row
after it is full, though the current animation needs work as it
happens too fast.

Removed some debug code for switching out the current piece type
and replaced it with a single debug command that will cycle
through all choices.