Clear a row when all squares are filled.
Made a board state that indicates if the board is in regular
moving state, in clearing state (when a row has been filled) or
in paused state (coming later).

Moved and improved the logic to tell when a row is filled and
implemented a cycle where the filled row is cleard, but the
squares above have not yet fallen to close the gap.