Refactor the movement code path.
Made a new function to check if a given shape and rotation
at a given x,y location is a valid placement.

Removed a bunch of code that was specific to every direction and
piece and now we just use the general purpose function.
What was that line about weeks of programming?