Changes to row clearing, score, board state
Updated how the board state changes when we clear a row.
Added basic scoring, currently just a point per cleared row.

Moved the logic for row clearing into the main event loop.
Added a function to place a new piece and check for end of game.
Added scoring whenever a row is cleared.