Show what piece is coming next
Draw a small grid to the left of the board and fill it with
the next tet in line to come off the stack.

Added more logic to place_new_piece to also update the next
array with the next piece in the queue.

Grouped the playfield and the next box code together.